ROX-35080: replace single key embed with bundle embed
Replace the single release-key-3.pub.txt embed with a bundle.json format
that supports multiple keys. Move ParseKeyBundle() and
BundleToSignatureIntegration() to pkg/signatures/ as public API, and
update central/signatureintegration/datastore to delegate to them.

Export DefaultRedHatIntegrationID and DefaultRedHatIntegrationName as
constants and update all callers to use them directly instead of
DefaultRedHatSignatureIntegration.GetId().

Partially generated by AI.
