+++ central/signatureintegration/datastore/singleton.go
@@ -47,7 +47,7 @@ func KeyBundleWatcher() Stoppable {
 func seedRedHatDefaultSignatureIntegration(siStore store.SignatureIntegrationStore) {
 	ctx := sac.WithGlobalAccessScopeChecker(context.Background(), sac.AllowAllAccessScopeChecker())
 
-	id := signatures.DefaultRedHatSignatureIntegration.GetId()
+	id := signatures.DefaultRedHatIntegrationID
 	_, exists, err := siStore.Get(ctx, id)
 	if err != nil {
 		utils.Should(errors.Wrap(err, "checking for default Red Hat signature integration"))
