+++ central/signatureintegration/datastore/singleton_test.go
@@ -14,7 +14,7 @@ func TestSeedFirstInstall(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	mockStore := storeMocks.NewMockSignatureIntegrationStore(ctrl)
 
-	id := signatures.DefaultRedHatSignatureIntegration.GetId()
+	id := signatures.DefaultRedHatIntegrationID
 	mockStore.EXPECT().Get(gomock.Any(), id).Return(nil, false, nil).Times(1)
 	mockStore.EXPECT().Upsert(gomock.Any(), signatures.DefaultRedHatSignatureIntegration).Return(nil).Times(1)
 
@@ -25,7 +25,7 @@ func TestSeedSubsequentStartup(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	mockStore := storeMocks.NewMockSignatureIntegrationStore(ctrl)
 
-	id := signatures.DefaultRedHatSignatureIntegration.GetId()
+	id := signatures.DefaultRedHatIntegrationID
 	mockStore.EXPECT().Get(gomock.Any(), id).Return(signatures.DefaultRedHatSignatureIntegration, true, nil).Times(1)
 
 	seedRedHatDefaultSignatureIntegration(mockStore)
